Fix floatertables such as the grader table
The grader table comes with floating headers that are altered through
javascript. The script seaches for a <header>-element or an element with
the class navbar, that has position: fixed. However, we work with
sticky. The two additional menu elements are not visible. The second one
serves as a dummy element from which the floater-js can get the menu
height. The first one just blocks the view to the grader table, which
would be visible in the small window right of the menu and on top of the
floating header.