Fix action menu on quiz edit page
On the tab ‘Question’ of the quiz edit page there is an ‘Add’ action
menu that allows to add questions to the quiz. However, the dropdown
menu did not properly show and was hidden in a containing
element. This meant that the user had to scroll down first to click on
items from the menu. This was cumbersome. This change adds additional
padding to the containing ul element such that the dropdown shows up
without scrolling.